Cleanup and restructure
