+++ .github/workflows/build.yaml
@@ -15,18 +15,18 @@ jobs:
           - os: macos-latest
             TARGET: macos
             CMD_BUILD: >
-                pyinstaller -F -w -n ledeffect-simulator -p .. simulator/simulator.py &&
+                pyinstaller -F -w -n ledeffect-simulator -p .. app.py &&
                 cd dist/ &&
                 zip -r9 ledeffect-simulator ledeffect-simulator.app
             OUT_FILE_NAME: ledeffect-simulator.zip
             ASSET_MIME: application/zip
-            POETRY_FOLDER: ~/.local/
+            POETRY_FOLDER: ~/.local
           - os: windows-latest
             TARGET: windows
-            CMD_BUILD: pyinstaller -F -w -n ledeffect-simulator -p .. simulator/simulator.py
+            CMD_BUILD: pyinstaller -F -w -n ledeffect-simulator -p .. app.py
             OUT_FILE_NAME: ledeffect-simulator.exe
             ASSET_MIME: application/vnd.microsoft.portable-executable
-            POETRY_FOLDER: C:\Users\runneradmin\.local\
+            POETRY_FOLDER: C:\Users\runneradmin\.local
     steps:
     - uses: actions/checkout@v1
     - name: Set up Python 3.9
